webapp/requirements.txt
Flask
Pillow
Numpy
